Try testing ThermoIce0 (it doesn't pass yet).
