+++ physics/src/modules/ThermoIce0.cpp
@@ -7,7 +7,7 @@
 
 #include "include/ThermoIce0.hpp"
 
-#include "include/IceGrowth.hpp"
+#include "include/IceMinima.hpp"
 #include "include/ModelArray.hpp"
 #include "include/NZLevels.hpp"
 #include "include/constants.hpp"
@@ -139,7 +139,7 @@ void ThermoIce0::calculateElement(size_t i, const TimestepTime& tst)
     }
 
     // Melt all ice if it is below minimum threshold
-    if (0. < hice[i] && hice[i] < IceGrowth::minimumIceThickness()) {
+    if (0. < hice[i] && hice[i] < IceMinima::h()) {
         if (deltaHi[i] < 0) {
             double scaling = oldHi[i] / deltaHi[i];
             topMelt[i] *= scaling;
